Reorganized utils.py
