+++ projetos/ReCycleGAN/src/utils/__init__.py
@@ -2,4 +2,5 @@
 
 from .data_transform import Images
 from .utils import *
+from .run import *
 from .data_loader import ImageDataset, get_img_dataloader
