+++ projetos/ReCycleGAN/src/utils/utils.py
@@ -1,8 +1,6 @@
 # pylint: disable=invalid-name
-"""Assorted functions."""
+"""File handling, image manipulation and memory query functions."""
 
-import gc
-import time
 import subprocess
 import json
 from pathlib import Path
@@ -14,7 +12,6 @@ from PIL import Image
 import torch
 from torchvision import transforms
 from torchvision.utils import make_grid
-import wandb
 import pynvml
 
 class Constants:
@@ -23,6 +20,7 @@ class Constants:
     WB_PROJECT = "cyclegan"
     WB_DB_UPLOAD_JOB = "dataset_upload"
     WB_DB_ARTIFACT_TYPE = "datasets"
+
 
 def remove_all_files(folder_path):
     """Remove all files in a folder."""
@@ -33,6 +31,7 @@ def remove_all_files(folder_path):
                 file.unlink()
             elif file.is_dir():
                 remove_all_files(file)
+
 
 def filter_dataframe(df, filter_dict):
     """Filter a DataFrame by multiple columns.
@@ -48,6 +47,7 @@ def filter_dataframe(df, filter_dict):
     for col, values in filter_dict.items():
         mask &= df[col].isin(values)
     return df[mask]
+
 
 def img_size_count(img_folder, extension='jpg', verbose=False):
     """Count the number of images by size in a folder.
@@ -114,55 +114,66 @@ def resize_and_crop(image_path, output_path, target_size, size_filter=None):
     """
     with Image.open(image_path) as img:
 
-        original_width, original_height = img.size
-        target_width, target_height = target_size
+        def _remove_black_borders(img, target_width, target_height):
+            gray_img = img.convert('L')
+            np_gray = np.array(gray_img)
+            mask = np_gray > 10
+            coords = np.argwhere(mask)
+            if coords.size > 0:
+                x0, y0 = coords.min(axis=0)
+                x1, y1 = coords.max(axis=0) + 1
 
+                if (x1-x0<target_width) or (y1-y0<target_height):
+                    return False
+
+                bbox = (y0, x0, y1, x1)
+                return img.crop(bbox)
+            else:
+                return None
+
+        def _resize_and_crop(img, target_width, target_height):
+            original_width, original_height = img.size
+            original_aspect = original_width / original_height
+            target_aspect = target_width / target_height
+
+            if original_aspect > target_aspect:
+                new_height = target_height
+                new_width = int(new_height * original_aspect)
+            else:
+                new_width = target_width
+                new_height = int(new_width / original_aspect)
+
+            left = (new_width - target_width) / 2
+            top = (new_height - target_height) / 2
+            right = left + target_width
+            bottom = target_height
+
+            img = img.resize((new_width, new_height), Image.LANCZOS) # pylint: disable=no-member
+            img = img.crop((left, top, right, bottom))
+            return img
+
+        original_width, original_height = img.size
         if size_filter is not None:
             if (original_width, original_height) not in size_filter:
                 return False
 
-        gray_img = img.convert('L')
-        np_gray = np.array(gray_img)
-        mask = np_gray > 10
-        coords = np.argwhere(mask)
-        if coords.size > 0:
-            x0, y0 = coords.min(axis=0)
-            x1, y1 = coords.max(axis=0) + 1
-
-            if (x1-x0<target_width) or (y1-y0<target_height):
-                return False
-
-            bbox = (y0, x0, y1, x1)
-            img = img.crop(bbox)
-            original_width, original_height = img.size
-        else:
+        target_width, target_height = target_size
+        img = _remove_black_borders(img, target_width, target_height)
+        if img is None:
             return False
 
-
-        original_aspect = original_width / original_height
-        target_aspect = target_width / target_height
-
-        if original_aspect > target_aspect:
-            new_height = target_height
-            new_width = int(new_height * original_aspect)
-        else:
-            new_width = target_width
-            new_height = int(new_width / original_aspect)
-
-        left = (new_width - target_width) / 2
-        top = (new_height - target_height) / 2
-        right = left + target_width
-        bottom = target_height
-
-        img = img.resize((new_width, new_height), Image.LANCZOS) # pylint: disable=no-member
-        img = img.crop((left, top, right, bottom))
+        img = _resize_and_crop(img, target_width, target_height)
         img.save(output_path)
 
         return True
     return False
 
-def show_img(img, title=None, figsize=(4, 3), show=False, change_scale=False, nrow=None, labels=None):
-    """Show an image using matplotlib.
+
+def show_img(img, title=None, figsize=(4, 3), nrow=None, labels=None):
+    """Show a group of images using matplotlib.
+
+    If negative values are present in the image tensor, it is assumed
+    the scale is [-1, 1], and is then changed to [0, 1].
 
     Attributes:
     ------------
@@ -173,12 +184,6 @@ def show_img(img, title=None, figsize=(4, 3), show=False, change_scale=False, nr
     figsize: tuple
         Figure size (width, height).
         (Default: (4, 3))
-    show: bool
-        Whether to display the image.
-        (Default: False)
-    change_scale: bool
-        Whether to change the scale of the image
-        from [-1, 1] to [0, 1].
     nrow: int
         Number of images per row to display if the image is a tensor.
         If None, the number of rows is calculated based on the
@@ -189,20 +194,20 @@ def show_img(img, title=None, figsize=(4, 3), show=False, change_scale=False, nr
         If None, no labels are displayed.
         (Default: None)
     """
+    if len(img.shape) > 4:
+        msg = 'Image tensor has more than 4 dimensions.'
+        raise ValueError(msg)
+    if img.min() < 0:
+        img = (img + 1) / 2
 
     width, height = figsize
     f_size = int(min(width, height) * 1.5)
 
-    if change_scale:
-        img = (img + 1) / 2
-    if len(img.shape) > 4:
-        msg = 'Image tensor has more than 4 dimensions.'
-        raise ValueError(msg)
     if len(img.shape) == 4:
         if nrow is None:
             nrow = int(max(4, min(8, np.ceil(img.shape[0] / 2))))
         grid = make_grid(img, nrow=nrow, normalize=False, scale_each=False)
-        return show_img(grid, title=title, figsize=figsize, show=show, labels=labels)
+        return show_img(grid, title=title, figsize=figsize, labels=labels)
 
     img = img.permute(1, 2, 0)
     if img.shape[2]==1:
@@ -225,10 +230,8 @@ def show_img(img, title=None, figsize=(4, 3), show=False, change_scale=False, nr
         axs.set_yticks(ticks=y_ticks, labels=y_lab, rotation=90, fontsize=f_size)
 
     plt.tight_layout()
-    if show:
-        plt.show()
-        return None
     return fig
+
 
 def image_folder_to_tensor(img_dir, img_size=(256, 256), img_glob='*'):
     """
@@ -265,160 +268,8 @@ def image_folder_to_tensor(img_dir, img_size=(256, 256), img_glob='*'):
     all_images_tensor = torch.stack(image_tensors)
     return all_images_tensor
 
-def save_model(model, local_path='model.pth', wandb_log=True):
-    """
-    Saves the model state to a local file and optionally logs it to Weights & Biases (WandB).
 
-    Saves a checkpoint of the model's state dictionary to the specified local file.
-    If `wandb_log` is True, the model will also be saved to WandB for remote logging.
-
-    Parameters:
-    ------------
-    model: torch.nn.Module
-        Model instance to save.
-    local_path: str
-        File path where the model will be saved locally.
-        (default: 'model.pth')
-    wandb_log: bool
-        Whether to log the model to WandB for version control
-        and experiment tracking.
-        (default: True)
-    """
-    # Save locally
-    torch.save(model.state_dict(), local_path)
-
-    # Save to WandB
-    if wandb_log:
-        wandb.save(local_path)
-
-def save_losses(loss_G, loss_D_A, loss_D_B, loss_G_ad, loss_G_cycle, loss_G_id, loss_G_plp, filename='losses.txt'):
-    """
-    Saves the generator and discriminator losses to a text file.
-
-    Saves a text file containing the losses for the generator and discriminators
-    (A and B) over the training epochs.
-
-    Args:
-    - loss_G (list): List of generator losses over the training epochs.
-    - loss_D_A (list): List of discriminator A losses over the training epochs.
-    - loss_D_B (list): List of discriminator B losses over the training epochs.
-    - filename (str): The file path where the losses will be saved. Defaults to 'losses.txt'.
-    """
-    np.savetxt(
-        filename,
-        np.column_stack((loss_G, loss_D_A, loss_D_B, loss_G_ad, loss_G_cycle, loss_G_id, loss_G_plp)),
-        header='Generator total loss, Discriminator A loss, Discriminator B loss')
-
-def train_one_epoch(epoch, model, train_A, train_B, device, n_samples=None, plp_step=0):
-    """
-    Trains the CycleGAN model for a single epoch and returns the generator and discriminator losses.
-
-    Args:
-    - epoch (int): The current epoch number.
-    - model (CycleGAN): The CycleGAN model instance.
-    - train_A (DataLoader): DataLoader for domain A training images.
-    - train_B (DataLoader): DataLoader for domain B training images.
-    - device (torch.device): The device on which the model and data are
-    loaded (e.g., 'cuda' or 'cpu').
-    - n_samples (int): Number of samples to train on per batch.
-    If None, train on all samples. Default is None.
-    - plp_step: Steps between Path Length Penalty calculations. Used to adjust
-    PLP loss value. Default is 0.
-
-    Returns:
-    - loss_G (float): The total loss of the generator for this epoch.
-    - loss_D_A (float): The total loss of discriminator A for this epoch.
-    - loss_D_B (float): The total loss of discriminator B for this epoch.
-
-    During training:
-    - It iterates through the batches of both domains (A and B) and performs
-    optimization on the generators and discriminators.
-    - Progress is tracked with a `tqdm` progress bar that shows current generator
-    and discriminator losses.
-    """
-    time_start = time.time()
-    progress_bar = tqdm(zip(train_A, train_B), desc=f'Epoch {epoch:03d}',
-                        leave=False, disable=False)
-
-    loss_G, loss_D_A, loss_D_B, loss_G_ad, loss_G_cycle, loss_G_id, loss_G_plp = 0, 0, 0, 0, 0, 0, 0
-    for batch_A, batch_B in progress_bar:
-        progress_bar.set_description(f'Epoch {epoch:03d}')
-
-        if n_samples is not None:
-            batch_A = batch_A[:n_samples]
-            batch_B = batch_B[:n_samples]
-
-        real_A = batch_A.to(device)
-        real_B = batch_B.to(device)
-
-        # Perform one optimization step
-        loss = model.optimize(real_A, real_B)
-        loss_G += loss.loss_G.item()
-        loss_D_A += loss.loss_D_A.item()
-        loss_D_B += loss.loss_D_B.item()
-        loss_G_ad += loss.loss_G_ad.item()
-        loss_G_cycle += loss.loss_G_cycle.item()
-        loss_G_id += loss.loss_G_id.item()
-        loss_G_plp += loss.loss_G_plp.item()
-
-        progress_bar.set_postfix({
-            'G_loss': f'{loss.loss_G.item():.4f}',
-            'D_A_loss': f'{loss.loss_D_A.item():.4f}',
-            'D_B_loss': f'{loss.loss_D_B.item():.4f}',
-            'GPU': f'{print_gpu_memory_usage("",True):.2f}%',
-        })
-
-        torch.cuda.empty_cache()
-        gc.collect()
-
-    progress_bar.close()
-
-    loss_G /= (len(train_A) + len(train_B)) / 2
-    loss_D_A /= len(train_A)
-    loss_D_B /= len(train_B)
-    loss_G_ad /= (len(train_A) + len(train_B)) / 2
-    loss_G_cycle /= (len(train_A) + len(train_B)) / 2
-    loss_G_id /= (len(train_A) + len(train_B)) / 2
-    loss_G_plp /= (len(train_A) + len(train_B)) / 2 * plp_step
-
-    msg = f'Epoch {epoch:03d}: G_loss={loss_G:.4g}, '
-    msg += f'D_A_loss={loss_D_A:.4g}, D_B_loss={loss_D_B:.4g}, '
-    msg += f'G_ad={loss_G_ad:.4g}, G_cycle={loss_G_cycle:.4g}, '
-    msg += f'G_id={loss_G_id:.4g}, G_plp={loss_G_plp:.4g}, '
-    msg += f'Time={time.time() - time_start:.2f} s'
-    print(msg)
-    return loss_G, loss_D_A, loss_D_B, loss_G_ad, loss_G_cycle, loss_G_id, loss_G_plp
-
-def plot_losses(train_losses, val_losses):
-    """
-    Plots the training and validation losses over the epochs.
-
-    Args:
-    - train_losses (list): List of training losses (e.g., generator losses) over the epochs.
-    - val_losses (list): List of validation losses over the epochs.
-
-    Displays:
-    - A line plot showing the progression of training and validation losses.
-    - Training and validation losses are plotted against the number of epochs.
-    """
-    plt.plot(
-        range(1, len(train_losses) + 1),
-        train_losses,
-        label='Training Loss',
-        linewidth=2, alpha=0.7)
-    plt.plot(
-        range(1, len(val_losses) + 1),
-        val_losses,
-        label='Validation Loss',
-        linewidth=2, alpha=0.7)
-    plt.title('CycleGAN Training Losses')
-    plt.xlabel('Epoch')
-    plt.ylabel('Loss')
-    plt.legend()
-    plt.show()
-
-
-def get_gpu_memory_usage():
+def get_gpu_memory_usage_dict():
     """Get list of dict with memory usage of all GPUs."""
     pynvml.nvmlInit()
     device_count = pynvml.nvmlDeviceGetCount()
@@ -437,7 +288,8 @@ def get_gpu_memory_usage():
     pynvml.nvmlShutdown()
     return gpu_memory_info
 
-def print_gpu_memory_usage(msg=None, short_msg=False):
+
+def get_gpu_memory_usage(msg=None, short_msg=False):
     """Print the memory usage of all GPUs.
 
     Attibutes:
@@ -452,28 +304,30 @@ def print_gpu_memory_usage(msg=None, short_msg=False):
         short_msg==True, returns the percentage of memory used.
         (default=False)
     """
-    gpu_memory_info = get_gpu_memory_usage()
+    gpu_memory_info = get_gpu_memory_usage_dict()
     if short_msg:
         if msg is None:
             msg = "GPU Memory Usage"
         total = sum(info['total_memory'] for info in gpu_memory_info)
         used = sum(info['used_memory'] for info in gpu_memory_info)
         if len(msg):
-            print(f"{msg}: {used / (1024 ** 2):.2f} MB ({used / total * 100:.2f}% used)")
+            return f"{msg}: {used / (1024 ** 2):.2f} MB ({used / total * 100:.2f}% used)"
         else:
-            return used / total * 100
+            return f'{used / total * 100:.2f}%'
 
+    out = ""
     ident = ""
     if msg is not None:
-        print(msg)
+        out = f'{msg}\n'
         ident = " " * 2
     if len(gpu_memory_info) == 0:
-        print(f"{ident}No GPUs found.")
+        out += f"{ident}No GPUs found."
     for info in gpu_memory_info:
-        print(f"{ident}GPU {info['gpu_index']}:")
-        print(f"{ident}  Total Memory: {info['total_memory'] / (1024 ** 2):.2f} MB")
-        print(f"{ident}  Used Memory: {info['used_memory'] / (1024 ** 2):.2f} MB")
-        print(f"{ident}  Free Memory: {info['free_memory'] / (1024 ** 2):.2f} MB")
+        out += f"{ident}GPU {info['gpu_index']}:\n"
+        out += f"{ident}  Total Memory: {info['total_memory'] / (1024 ** 2):.2f} MB\n"
+        out += f"{ident}  Used Memory: {info['used_memory'] / (1024 ** 2):.2f} MB\n"
+        out += f"{ident}  Free Memory: {info['free_memory'] / (1024 ** 2):.2f} MB"
+    return out
 
 
 def count_parameters(model: torch.nn.Module) -> int:
@@ -491,14 +345,16 @@ def count_parameters(model: torch.nn.Module) -> int:
     """
     return sum(p.numel() for p in model.parameters())
 
+
 def get_current_commit():
     """Get current git hash of the repository."""
     try:
-        git_hash = subprocess.check_output(['git', 'rev-parse', 'HEAD']).strip().decode('utf-8')
-        commit_message = subprocess.check_output(['git', 'log', '-1', '--pretty=%B']).strip().decode('utf-8')
-        return git_hash, commit_message
+        git_hash = subprocess.check_output(['git', 'rev-parse', 'HEAD'])
+        commit_message = subprocess.check_output(['git', 'log', '-1', '--pretty=%B'])
+        return git_hash.strip().decode('utf-8'), commit_message.strip().decode('utf-8')
     except subprocess.CalledProcessError:
         return None, None
+
 
 def save_dict_as_json(data, file_path):
     """
